Create pages. Create and populate 2 directories
Page 'index.html' was created, to serve as a homepage.

Directory 'recipes' was created to host the pages with recipes.

Page 'recipes/chakalaka.html' was created to hold the first recipe.

A new directory ('images') was created to store the images for the recipes.
It was populated with three pictures"
	images/chakalaka.jpg
    images/it_chkn_mar.jpg
    images/teriyaki.jpg